+++ index.html
@@ -25,9 +25,8 @@
     <div class="card text-center">
       <div class="card-header">Pick your palette!</div>
       <div class="card-body">
-        <div id="button-panel"></div>
         <h5 class="card-title">Choose your colors</h5>
-        <div id="color-button"></div>
+        <div id="button-panel"></div>
       </div>
       <div class="card-footer">
         Card footer. Save button here maybe?
@@ -38,14 +37,12 @@
     </div>
   </div>
   <script
-    src="https://unpkg.com/react@16/umd/react.production.min.js"
+    src="https://unpkg.com/react@16/umd/react.development.js"
     crossorigin
   ></script>
   <script
-    src="https://unpkg.com/react-dom@16/umd/react-dom.production.min.js"
+    src="https://unpkg.com/react-dom@16/umd/react-dom.development.js"
     crossorigin
   ></script>
-  <script src="https://unpkg.com/babel-standalone@6/babel.min.js"></script>
-  <script src="./src/ReactComponents/ButtonPanel.js"></script>
-  <script src="./src/ReactComponents/ColorSelectButton.js"></script>
+  <script src="ButtonPanel.js"></script>
 </body>
