--- package.json
@@ -1,24 +1,0 @@
-{
-  "name": "palettepicker",
-  "version": "1.0.0",
-  "description": "This tool is a page that shows a generic/example web page in a preview window and allows the user to select colors for the different HTML elements on that example page. The aim of this tool is to help users choose a desired palette of colors for use on their own website layout, and give the user the hex or rgb values of those colors to be easily plugged into CSS.",
-  "main": "index.js",
-  "scripts": {
-    "test": "echo \"Error: no test specified\" && exit 1"
-  },
-  "repository": {
-    "type": "git",
-    "url": "git+https://github.com/nashvillefcc/palettepicker.git"
-  },
-  "keywords": [],
-  "author": "",
-  "license": "ISC",
-  "bugs": {
-    "url": "https://github.com/nashvillefcc/palettepicker/issues"
-  },
-  "homepage": "https://github.com/nashvillefcc/palettepicker#readme",
-  "dependencies": {
-    "babel-cli": "^6.26.0",
-    "babel-preset-react-app": "^3.1.2"
-  }
-}
